Fixed 'Bug 60622 - Can't get syntax highlighting/colorization for Xamarin.iOS bindings'.
